refactor(KERN-1560): split sql-to-meerkat.ts into step modules
Extract pipeline steps into focused files under steps/:
- steps/parse-ast.ts: SQL parsing + AST validation (UNION/RECURSIVE rejection)
- steps/classify-expressions.ts: measure/dimension classification + batch serialization
- steps/extract-modifiers.ts: ORDER BY, LIMIT, OFFSET extraction

sql-to-meerkat.ts is now a slim orchestrator (150 lines) that coordinates
the steps. Each step is independently testable.

work-item: KERN-1560
