+++ meerkat-core/src/sql-to-meerkat/steps/index.ts
@@ -1,0 +1,6 @@
+export { parseAndValidateAst } from './parse-ast';
+export type { ParseResult } from './parse-ast';
+export { classifyExpressions } from './classify-expressions';
+export type { ClassificationResult } from './classify-expressions';
+export { extractModifiers } from './extract-modifiers';
+export type { ModifierResult } from './extract-modifiers';
